refactor: centralize image helpers
